services/server/requirements.txt
requests
websockets
pydantic
python-dotenv
colorlog
fastapi[all]
nats-py==2.0.0a2
asyncio==3.4.3
numpy==1.26.4
pandas==1.3.5